Move HTTPRequest into own swift file
