+++ Sources/CorePayments/Networking/HTTP.swift
@@ -31,12 +31,3 @@ class HTTP {
         return HTTPResponse(status: response.statusCode, body: data)
     }
 }
-
-/// :nodoc:
-public struct HTTPRequest {
-    
-    let url: URL
-    let method: HTTPMethod
-    let body: Data?
-    let headers: [HTTPHeader: String]
-}
